postgres_to_redis/requirements.txt
pydantic==1.10.7
redis==5.0.0
psycopg2-binary==2.9